Add files via upload
This Word Counter is a Graphical user interface(GUI) system that allows user to easily count the words written in TextArea using button. Users may very quickly count the words using the system simple design. The Word Counter has simple design with button and TextArea. The user can write a text in TextArea click on the button to count the words.

The users may count the words written in Text Area using count buttons. The user can write their content or sentences in the Text Area and after clicking the count button the numbers of words in the Text Area will be displayed to the users. Further user can change the text and can again count the words.

The Java Swing library which gives a flexible set of components is used to build the Word Counter. The application makes use of Swing components such as JTextArea, JButton, JLabel and JPanel among many other Swing components. Button event is handled by the ActionListener interface and button events are defined using ActionListener’s actionPerformed() method.

Methods used in the project:

actionPerformed():

The action or events of the buttons are handled using this method. Buttons in the system to count the number of words in TextArea is handled using this method. This method describe how the component should behave after the button is click.

split():

This method is used to split the text based on white-space. This method will check for the white-space and it will split text into words and further counts the number of words in the Text Area.